+++ client/src/app/dataset-table/components/dataset-table.component.ts
@@ -1,5 +1,4 @@
 import { Component } from '@angular/core';
-import { Dataset } from 'shared/dtos/dataset.dto';
 import { DatasetService } from '../../core/services/dataset.service';
 
 /**
@@ -9,25 +8,20 @@ import { DatasetService } from '../../core/services/dataset.service';
 @Component({
   selector: 'dataset-table',
   template: `
-    <mat-expansion-panel *ngFor="let dataset of datasets">
-      <mat-expansion-panel-header>
-        <mat-panel-title>{{ dataset.name }}</mat-panel-title>
-        <mat-panel-description>{{ dataset.description }}</mat-panel-description>
-      </mat-expansion-panel-header>
+    <div *ngIf="datasetService.datasets | async as datasets">
+      <mat-expansion-panel *ngFor="let dataset of datasets">
+        <mat-expansion-panel-header>
+          <mat-panel-title>{{ dataset.name }}</mat-panel-title>
+          <mat-panel-description>{{
+            dataset.description
+          }}</mat-panel-description>
+        </mat-expansion-panel-header>
 
-      <entry-table [dataset]="dataset"></entry-table>
-    </mat-expansion-panel>
+        <entry-table [dataset]="dataset"></entry-table>
+      </mat-expansion-panel>
+    </div>
   `,
 })
 export class DatasetTable {
-  /** All of the available datasets */
-  datasets: Dataset[] = [];
-
-  constructor(private datasetService: DatasetService) {
-    this.loadDatasets();
-  }
-
-  async loadDatasets(): Promise<void> {
-    this.datasets = await this.datasetService.getDatasets();
-  }
+  constructor(public datasetService: DatasetService) {}
 }