+++ client/src/app/datasets/datasets-routing.module.ts
@@ -3,9 +3,11 @@ import { RouterModule, Routes } from '@angular/router';
 
 // Components
 import { DatasetControlComponent } from './components/dataset-control.component';
+import { ProjectAccess } from './components/project-access.component';
 
 const routes: Routes = [
   { path: 'dataset-control', component: DatasetControlComponent },
+  { path: 'project-access', component: ProjectAccess },
 ];
 
 @NgModule({