+++ server/src/dataset/dataset.service.ts
@@ -2,6 +2,8 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Dataset } from './dataset.schema';
 import { Injectable } from '@nestjs/common';
 import { FilterQuery, Model } from 'mongoose';
+import { ProjectAccessChangeFull } from './dataset.dto';
+import { Project } from '../project/project.schema';
 
 @Injectable()
 export class DatasetService {
@@ -30,5 +32,34 @@ export class DatasetService {
     dataset: Pick<Dataset, Exclude<keyof Dataset, '_id'>>,
   ): Promise<Dataset> {
     return this.datasetModel.create(dataset);
+  }
+
+  /**
+   * Check to see if a dataset exists
+   */
+  async exists(name: string): Promise<boolean> {
+    return (await this.datasetModel.findOne({ name }).exec()) !== null;
+  }
+
+  async getByProject(project: Project): Promise<Dataset[]> {
+    return this.datasetModel
+      .find({ [`projectAccess.${project._id}`]: true })
+      .exec();
+  }
+
+  async changeProjectAccess(
+    projectAccessChange: ProjectAccessChangeFull,
+  ): Promise<void> {
+    await this.datasetModel
+      .updateOne(
+        { _id: projectAccessChange.dataset._id },
+        {
+          $set: {
+            [`projectAccess.${projectAccessChange.project._id}`]:
+              projectAccessChange.hasAccess,
+          },
+        },
+      )
+      .exec();
   }
 }